pyproject.toml
[project]
name = "lory"
version = "0.1.0"
description = ""
authors = [
    {name = "JuanPinheiroFIAP",email = "rm552202@fiap.com.br"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)"
]

[tool.poetry]
packages = [{include = "lory", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "ruff (>=0.15.4,<0.16.0)",
    "detect-secrets (>=1.5.0,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
